+++ src/Time/StepChoosers/PreventRapidIncrease.hpp
@@ -9,10 +9,10 @@
 
 #include "Options/String.hpp"
 #include "Time/History.hpp"
+#include "Time/SlabRoundingError.hpp"
 #include "Time/StepChoosers/StepChooser.hpp"
 #include "Time/Tags/HistoryEvolvedVariables.hpp"
 #include "Time/TimeStepRequest.hpp"
-#include "Time/Utilities.hpp"
 #include "Utilities/Serialization/CharmPupable.hpp"
 #include "Utilities/TMPL.hpp"
